README
This file holds a wealth of knowledge on how to effectively use and personalize this website to your liking. Within its contents, you'll find valuable insights and practical guidance to make it truly your own. With each section, you'll gain a deeper understanding of its features and functionalities, empowering you to navigate and customize with ease